- переделал тип поля salary на BigDecimal

+++ src/main/java/edu/skypro/homework/controller/EmployeeController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 
 @RestController
@@ -21,7 +22,7 @@ public class EmployeeController {
     @GetMapping(path = "/add")
     public Employee add(@RequestParam String firstName,
                         @RequestParam String lastName,
-                        @RequestParam double salary,
+                        @RequestParam BigDecimal salary,
                         @RequestParam int department) {
         return emplService.add(firstName, lastName, salary, department);
     }
@@ -29,7 +30,7 @@ public class EmployeeController {
     @GetMapping(path = "/remove")
     public Employee remove(@RequestParam String firstName,
                            @RequestParam String lastName,
-                           @RequestParam double salary,
+                           @RequestParam BigDecimal salary,
                            @RequestParam int department) {
         return emplService.remove(firstName, lastName, salary, department);
     }
@@ -37,7 +38,7 @@ public class EmployeeController {
     @GetMapping(path = "/find")
     public Employee find(@RequestParam String firstName,
                          @RequestParam String lastName,
-                         @RequestParam double salary,
+                         @RequestParam BigDecimal salary,
                          @RequestParam int department) {
         return emplService.find(firstName, lastName, salary, department);
     }
